+++ x/community/types/expected_keepers.go
@@ -29,4 +29,5 @@ type HardKeeper interface {
 type DistributionKeeper interface {
 	DistributeFromFeePool(ctx sdk.Context, amount sdk.Coins, receiveAddr sdk.AccAddress) error
 	FundCommunityPool(ctx sdk.Context, amount sdk.Coins, sender sdk.AccAddress) error
+	GetFeePoolCommunityCoins(ctx sdk.Context) sdk.DecCoins
 }