+++ CHANGELOG.md
@@ -44,6 +44,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (deps) [#1477] Migrate to CometBFT.
 - (x/incentive) [#1512] Add grpc query service.
 - (deps) [#1544] Bump confio/ics23/go to v0.9.0, cosmos/keyring to v1.2.0.
+- (x/community) [#1563] Include x/community module pool balance in x/distribution
+  community_pool query response.
 
 ### Deprecated
 
@@ -228,6 +230,8 @@ the [changelog](https://github.com/cosmos/cosmos-sdk/blob/v0.38.4/CHANGELOG.md).
 - [#257](https://github.com/Kava-Labs/kava/pulls/257) Include scripts to run
   large-scale simulations remotely using aws-batch
 
+
+[#1563]: https://github.com/Kava-Labs/kava/pull/1563
 [#1544]: https://github.com/Kava-Labs/kava/pull/1544
 [#1477]: https://github.com/Kava-Labs/kava/pull/1477
 [#1512]: https://github.com/Kava-Labs/kava/pull/1512